Fix merge: rv32i state integer size widths
This incompatibility resulted from the changes of
427b997d82d746029a6b390a77cabada8fae7d0e and
998405a2752ae70d9e6428616ead07f814665fc8. The Riscv32iStoredState is
changed to u32 types instead of usize, while the RISC-V CSRs are
extracted into a generic crate, irrespective of integer width. It is
safe to cast here, as usize is equivalent to u32 for `rv32i`
platforms.

+++ arch/rv32i/src/syscall.rs
@@ -140,7 +140,7 @@ impl kernel::syscall::UserspaceKernelBoundary for SysCall {
         _state: &mut Riscv32iStoredState,
     ) -> (*mut usize, ContextSwitchReason) {
         // Convince lint that 'mcause' and 'R_A4' are used during test build
-        let _cause = mcause::Trap::from(_state.mcause);
+        let _cause = mcause::Trap::from(_state.mcause as usize);
         let _arg4 = _state.regs[R_A4];
         unimplemented!()
     }
@@ -367,7 +367,7 @@ impl kernel::syscall::UserspaceKernelBoundary for SysCall {
           : "memory"
           : "volatile");
 
-        let ret = match mcause::Trap::from(state.mcause) {
+        let ret = match mcause::Trap::from(state.mcause as usize) {
             mcause::Trap::Interrupt(_intr) => {
                 // An interrupt occurred while the app was running.
                 ContextSwitchReason::Interrupted
@@ -468,7 +468,7 @@ impl kernel::syscall::UserspaceKernelBoundary for SysCall {
             stack_pointer as usize,
             state.mcause,
         ));
-        crate::print_mcause(mcause::Trap::from(state.mcause), writer);
+        crate::print_mcause(mcause::Trap::from(state.mcause as usize), writer);
         let _ = writer.write_fmt(format_args!(
             ")\
              \r\n mtval:  {:#010X}\
